+++ CHANGELOG.md
@@ -2,6 +2,10 @@
  
 ## Unreleased
 
+### Features
+
+- Add sendEmailVerification and verifyEmail
+
 ## v9.5.0 (08/01/2025)
 
 ### Changes